requirements.txt
requests
python-dotenv
moviepy>=2.0.0
Pillow
numpy
